+++ cloudwatch_writer_test.go
@@ -2,7 +2,7 @@ package zerolog2cloudwatch_test
 
 import (
 	"encoding/json"
-	"errors"
+	"fmt"
 	"io"
 	"sync"
 	"testing"
@@ -12,19 +12,21 @@ import (
 	"github.com/aws/aws-sdk-go/aws/awserr"
 	"github.com/aws/aws-sdk-go/service/cloudwatchlogs"
 	"github.com/mec07/zerolog2cloudwatch"
+	"github.com/pkg/errors"
 	"github.com/stretchr/testify/assert"
 )
 
 type mockClient struct {
-	sync.Mutex
-	logEvents     []*cloudwatchlogs.InputLogEvent
-	logGroupName  *string
-	logStreamName *string
+	sync.RWMutex
+	putLogEventsShouldError bool
+	logEvents               []*cloudwatchlogs.InputLogEvent
+	logGroupName            *string
+	logStreamName           *string
 }
 
 func (c *mockClient) DescribeLogStreams(*cloudwatchlogs.DescribeLogStreamsInput) (*cloudwatchlogs.DescribeLogStreamsOutput, error) {
-	c.Lock()
-	defer c.Unlock()
+	c.RLock()
+	defer c.RUnlock()
 
 	if c.logGroupName == nil {
 		return nil, awserr.New(cloudwatchlogs.ErrCodeResourceNotFoundException, "blah", nil)
@@ -61,6 +63,9 @@ func (c *mockClient) CreateLogStream(input *cloudwatchlogs.CreateLogStreamInput)
 func (c *mockClient) PutLogEvents(putLogEvents *cloudwatchlogs.PutLogEventsInput) (*cloudwatchlogs.PutLogEventsOutput, error) {
 	c.Lock()
 	defer c.Unlock()
+	if c.putLogEventsShouldError {
+		return nil, errors.New("should error")
+	}
 
 	if putLogEvents == nil {
 		return nil, errors.New("received nil *cloudwatchlogs.PutLogEventsInput")
@@ -74,8 +79,8 @@ func (c *mockClient) PutLogEvents(putLogEvents *cloudwatchlogs.PutLogEventsInput
 }
 
 func (c *mockClient) getLogEvents() []*cloudwatchlogs.InputLogEvent {
-	c.Lock()
-	defer c.Unlock()
+	c.RLock()
+	defer c.RUnlock()
 
 	logEvents := make([]*cloudwatchlogs.InputLogEvent, len(c.logEvents))
 	copy(logEvents, c.logEvents)
@@ -83,9 +88,63 @@ func (c *mockClient) getLogEvents() []*cloudwatchlogs.InputLogEvent {
 	return logEvents
 }
 
+func (c *mockClient) waitForLogs(numberOfLogs int, timeout time.Duration) error {
+	endTime := time.Now().Add(timeout)
+	for {
+		if c.numLogs() >= numberOfLogs {
+			return nil
+		}
+
+		if time.Now().After(endTime) {
+			return errors.New("ran out of time waiting for logs")
+		}
+
+		time.Sleep(time.Millisecond)
+	}
+}
+
+func (c *mockClient) numLogs() int {
+	c.RLock()
+	defer c.RUnlock()
+
+	return len(c.logEvents)
+}
+
 type exampleLog struct {
 	Time, Message, Filename string
 	Port                    uint16
+}
+
+type logsContainer struct {
+	sync.Mutex
+	logs []exampleLog
+}
+
+func (l *logsContainer) addLog(log exampleLog) {
+	l.Lock()
+	defer l.Unlock()
+
+	l.logs = append(l.logs, log)
+}
+
+func (l *logsContainer) getLogEvents() ([]*cloudwatchlogs.InputLogEvent, error) {
+	l.Lock()
+	defer l.Unlock()
+
+	var logEvents []*cloudwatchlogs.InputLogEvent
+	for _, log := range l.logs {
+		message, err := json.Marshal(log)
+		if err != nil {
+			return nil, errors.Wrap(err, "json.Marshal")
+		}
+
+		logEvents = append(logEvents, &cloudwatchlogs.InputLogEvent{
+			Message: aws.String(string(message)),
+			// Timestamps for CloudWatch Logs should be in milliseconds since the epoch.
+			Timestamp: aws.Int64(time.Now().UTC().UnixNano() / int64(time.Millisecond)),
+		})
+	}
+	return logEvents, nil
 }
 
 func helperWriteLogs(t *testing.T, writer io.Writer, logs ...interface{}) {
@@ -99,22 +158,6 @@ func helperWriteLogs(t *testing.T, writer io.Writer, logs ...interface{}) {
 			t.Fatalf("writer.Write(%s): %v", string(message), err)
 		}
 	}
-}
-
-func helperMakeLogEvents(t *testing.T, logs ...interface{}) []*cloudwatchlogs.InputLogEvent {
-	var logEvents []*cloudwatchlogs.InputLogEvent
-	for _, log := range logs {
-		message, err := json.Marshal(log)
-		if err != nil {
-			t.Fatalf("json.Marshal: %v", err)
-		}
-		logEvents = append(logEvents, &cloudwatchlogs.InputLogEvent{
-			Message: aws.String(string(message)),
-			// Timestamps for CloudWatch Logs should be in milliseconds since the epoch.
-			Timestamp: aws.Int64(time.Now().UTC().UnixNano() / int64(time.Millisecond)),
-		})
-	}
-	return logEvents
 }
 
 // assertEqualLogMessages asserts that the log messages are all the same, ignoring the timestamps.
@@ -137,27 +180,239 @@ func assertEqualLogMessages(t *testing.T, expectedLogs []*cloudwatchlogs.InputLo
 func TestCloudWatchWriter(t *testing.T) {
 	client := &mockClient{}
 
-	cloudWatchWriter, err := zerolog2cloudwatch.NewWriterWithClient(client, "logGroup", "logStream")
+	cloudWatchWriter, err := zerolog2cloudwatch.NewWriterWithClient(client, 200*time.Millisecond, "logGroup", "logStream")
 	if err != nil {
 		t.Fatalf("NewWriterWithClient: %v", err)
 	}
+	defer cloudWatchWriter.Close()
 
-	aLog1 := exampleLog{
+	// give the queueMonitor goroutine time to start up
+	time.Sleep(time.Millisecond)
+
+	log1 := exampleLog{
 		Time:     "2009-11-10T23:00:02.043123061Z",
 		Message:  "Test message 1",
 		Filename: "filename",
 		Port:     666,
 	}
-	aLog2 := exampleLog{
+	log2 := exampleLog{
 		Time:     "2009-11-10T23:00:02.043123061Z",
 		Message:  "Test message 2",
 		Filename: "filename",
 		Port:     666,
 	}
+	log3 := exampleLog{
+		Time:     "2009-11-10T23:00:02.043123061Z",
+		Message:  "Test message 3",
+		Filename: "filename",
+		Port:     666,
+	}
+	log4 := exampleLog{
+		Time:     "2009-11-10T23:00:02.043123061Z",
+		Message:  "Test message 4",
+		Filename: "filename",
+		Port:     666,
+	}
 
-	helperWriteLogs(t, cloudWatchWriter, aLog1, aLog2)
+	helperWriteLogs(t, cloudWatchWriter, log1, log2, log3, log4)
 
-	expectedLogs := helperMakeLogEvents(t, aLog1, aLog2)
+	logs := logsContainer{}
+	logs.addLog(log1)
+	logs.addLog(log2)
+	logs.addLog(log3)
+	logs.addLog(log4)
+
+	expectedLogs, err := logs.getLogEvents()
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	if err = client.waitForLogs(2, 201*time.Millisecond); err != nil {
+		t.Fatal(err)
+	}
 
 	assertEqualLogMessages(t, expectedLogs, client.getLogEvents())
+}
+
+func TestCloudWatchWriterBatchInterval(t *testing.T) {
+	client := &mockClient{}
+
+	cloudWatchWriter, err := zerolog2cloudwatch.NewWriterWithClient(client, 5*time.Second, "logGroup", "logStream")
+	if err != nil {
+		t.Fatalf("NewWriterWithClient: %v", err)
+	}
+	defer cloudWatchWriter.Close()
+
+	// can't set anything below 200 milliseconds
+	err = cloudWatchWriter.SetBatchInterval(199 * time.Millisecond)
+	if err == nil {
+		t.Fatal("expected an error")
+	}
+
+	// setting it to a value greater than or equal to 200 is OK
+	err = cloudWatchWriter.SetBatchInterval(200 * time.Millisecond)
+	if err != nil {
+		t.Fatalf("CloudWatchWriter.SetBatchInterval: %v", err)
+	}
+
+	// give the queueMonitor goroutine time to start up
+	time.Sleep(time.Millisecond)
+
+	aLog := exampleLog{
+		Time:     "2009-11-10T23:00:02.043123061Z",
+		Message:  "Test message",
+		Filename: "filename",
+		Port:     666,
+	}
+
+	helperWriteLogs(t, cloudWatchWriter, aLog)
+
+	// The client shouldn't have received any logs at this time
+	assert.Equal(t, 0, client.numLogs())
+
+	// Still no logs after 100 milliseconds
+	time.Sleep(100 * time.Millisecond)
+	assert.Equal(t, 0, client.numLogs())
+
+	// The client should have received the log after another 101 milliseconds
+	// (as that is a total sleep time of 201 milliseconds)
+	time.Sleep(101 * time.Millisecond)
+	assert.Equal(t, 1, client.numLogs())
+}
+
+// Hit the 1MB log limit to trigger an earlier batch
+func TestCloudWatchWriterBulk(t *testing.T) {
+	client := &mockClient{}
+
+	cloudWatchWriter, err := zerolog2cloudwatch.NewWriterWithClient(client, 200*time.Millisecond, "logGroup", "logStream")
+	if err != nil {
+		t.Fatalf("NewWriterWithClient: %v", err)
+	}
+	defer cloudWatchWriter.Close()
+
+	logs := logsContainer{}
+	numLogs := 10000
+	for i := 0; i < numLogs; i++ {
+		aLog := exampleLog{
+			Time:     "2009-11-10T23:00:02.043123061Z",
+			Message:  fmt.Sprintf("Test message %d", i),
+			Filename: "filename",
+			Port:     666,
+		}
+		helperWriteLogs(t, cloudWatchWriter, aLog)
+		logs.addLog(aLog)
+	}
+
+	// Main assertion is that we are triggering a batch early as we're sending so much data
+	assert.True(t, client.numLogs() > 0)
+
+	if err = client.waitForLogs(numLogs, 200*time.Millisecond); err != nil {
+		t.Fatal(err)
+	}
+
+	expectedLogs, err := logs.getLogEvents()
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	assertEqualLogMessages(t, expectedLogs, client.getLogEvents())
+}
+
+func TestCloudWatchWriterParallel(t *testing.T) {
+	client := &mockClient{}
+
+	cloudWatchWriter, err := zerolog2cloudwatch.NewWriterWithClient(client, 200*time.Millisecond, "logGroup", "logStream")
+	if err != nil {
+		t.Fatalf("NewWriterWithClient: %v", err)
+	}
+	defer cloudWatchWriter.Close()
+
+	logs := logsContainer{}
+	numLogs := 8000
+	for i := 0; i < numLogs; i++ {
+		go func() {
+			aLog := exampleLog{
+				Time:     "2009-11-10T23:00:02.043123061Z",
+				Message:  "Test message",
+				Filename: "filename",
+				Port:     666,
+			}
+			logs.addLog(aLog)
+			helperWriteLogs(t, cloudWatchWriter, aLog)
+		}()
+	}
+
+	// allow more time as there are a lot of goroutines to set off!
+	if err = client.waitForLogs(numLogs, 2*time.Second); err != nil {
+		t.Fatal(err)
+	}
+
+	expectedLogs, err := logs.getLogEvents()
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	assertEqualLogMessages(t, expectedLogs, client.getLogEvents())
+}
+
+func TestCloudWatchWriterClose(t *testing.T) {
+	client := &mockClient{}
+
+	cloudWatchWriter, err := zerolog2cloudwatch.NewWriterWithClient(client, 200*time.Millisecond, "logGroup", "logStream")
+	if err != nil {
+		t.Fatalf("NewWriterWithClient: %v", err)
+	}
+	defer cloudWatchWriter.Close()
+
+	numLogs := 99
+	for i := 0; i < numLogs; i++ {
+		aLog := exampleLog{
+			Time:     "2009-11-10T23:00:02.043123061Z",
+			Message:  fmt.Sprintf("Test message %d", i),
+			Filename: "filename",
+			Port:     666,
+		}
+		helperWriteLogs(t, cloudWatchWriter, aLog)
+	}
+
+	// The logs shouldn't have come through yet
+	assert.Equal(t, 0, client.numLogs())
+
+	// Close should block until the queue is empty
+	cloudWatchWriter.Close()
+
+	// The logs should have all come through now
+	assert.Equal(t, numLogs, client.numLogs())
+}
+
+func TestCloudWatchWriterReportError(t *testing.T) {
+	client := &mockClient{
+		putLogEventsShouldError: true,
+	}
+
+	cloudWatchWriter, err := zerolog2cloudwatch.NewWriterWithClient(client, 200*time.Millisecond, "logGroup", "logStream")
+	if err != nil {
+		t.Fatalf("NewWriterWithClient: %v", err)
+	}
+	defer cloudWatchWriter.Close()
+
+	// give the queueMonitor goroutine time to start up
+	time.Sleep(time.Millisecond)
+
+	log1 := exampleLog{
+		Time:     "2009-11-10T23:00:02.043123061Z",
+		Message:  "Test message 1",
+		Filename: "filename",
+		Port:     666,
+	}
+
+	helperWriteLogs(t, cloudWatchWriter, log1)
+
+	// sleep until the batch should have been sent
+	time.Sleep(201 * time.Millisecond)
+
+	_, err = cloudWatchWriter.Write([]byte("hello world"))
+	if err == nil {
+		t.Fatal("expected the last error from PutLogEvents to appear here")
+	}
 }